Fix compile error if SystemLinearizer was used with a Scalar other than double
In LinearSystem, the virtual methods getDerivativeControl and getDerivativeState are defined with time_t as their time parameter, which is a typedef for the template parameter SCALAR. The SystemLinearizer however uses a hardcoded double here. Thus, compilation fails if the class is instantiated with a different scalar type, as the override is no longer valid.