Fix compile error if SystemLinearizer was used with a Scalar other than double (#50)
* Fix compile error if SystemLinearizer was used with a Scalar other than double

In LinearSystem, the virtual methods getDerivativeControl and getDerivativeState are defined with time_t as their time parameter, which is a typedef for the template parameter SCALAR. The SystemLinearizer however uses a hardcoded double here. Thus, compilation fails if the class is instantiated with a different scalar type, as the override is no longer valid.

* Add typedef for time_t

Apparently, the typedef time_t from time.h or a similar C header was used instead of the typedef from the base class. This redefinition should fix that.